bridge: migrate apis to batch sdk v2
