+++ src/consts.ts
@@ -1,6 +1,7 @@
 export const Consts = {
   AttributeKeyRegexp: /^[a-zA-Z0-9_]{1,30}$/,
   AttributeStringMaxLength: 64,
+  AttributeStringArrayMaxSize: 25,
   BatchPluginName: "Batch",
   DevelopmentMode: true,
   EventDataStringMaxLength: 64,
