--- src/modules/user/eventData.ts
@@ -1,142 +1,0 @@
-import { BatchSDK } from "../../../types";
-import Consts from "../../consts";
-import { isBoolean, isNumber, isString, writeBatchLog } from "../../helpers";
-
-export enum TypedEventAttributeType {
-  String = "s",
-  Boolean = "b",
-  Integer = "i",
-  Float = "f",
-  Date = "d",
-  URL = "u",
-}
-
-export interface ITypedEventAttribute {
-  type: TypedEventAttributeType;
-  value: string | boolean | number;
-}
-
-export interface IEventDataInternalRepresentation {
-  tags: string[];
-  attributes: { [key: string]: ITypedEventAttribute };
-}
-
-export class BatchEventData implements BatchSDK.BatchEventData {
-  // Store the tags in a map, since Set isn't supported on Android 4.4
-  private _tags: { [key: string]: true };
-  private _attributes: { [key: string]: ITypedEventAttribute };
-
-  constructor() {
-    this._tags = {};
-    this._attributes = {};
-  }
-
-  public addTag(tag: string): BatchEventData {
-    if (typeof tag === "undefined") {
-      writeBatchLog(false, "BatchEventData - A tag is required");
-      return this;
-    }
-
-    if (isString(tag)) {
-      if (tag.length === 0 || tag.length > Consts.EventDataStringMaxLength) {
-        writeBatchLog(
-          false,
-          "BatchEventData - Tags can't be empty or longer than " +
-            Consts.EventDataStringMaxLength +
-            " characters. Ignoring tag '" +
-            tag +
-            "'."
-        );
-        return this;
-      }
-    } else {
-      writeBatchLog(false, "BatchEventData - Tag argument must be a string");
-      return this;
-    }
-
-    this._tags[tag.toLowerCase()] = true;
-
-    return this;
-  }
-
-  public put(
-    key: string,
-    value: string | number | boolean | Date | URL
-  ): BatchEventData {
-    if (!isString(key)) {
-      writeBatchLog(false, "BatchEventData - Key must be a string");
-      return this;
-    }
-
-    if (!Consts.AttributeKeyRegexp.test(key || "")) {
-      writeBatchLog(
-        false,
-        "BatchEventData - Invalid key. Please make sure that the key is made of letters, underscores and numbers only (a-zA-Z0-9_). It also can't be longer than 30 characters. Ignoring attribute '" +
-          key +
-          "'"
-      );
-      return this;
-    }
-
-    if (typeof value === "undefined" || value === null) {
-      writeBatchLog(
-        false,
-        "BatchEventData - Value cannot be undefined or null"
-      );
-      return this;
-    }
-
-    key = key.toLowerCase();
-
-    let typedAttrValue: ITypedEventAttribute | undefined;
-
-    if (value instanceof Date) {
-      typedAttrValue = {
-        type: TypedEventAttributeType.Date,
-        value: value.getTime(),
-      };
-    } else if (value instanceof URL) {
-      typedAttrValue = {
-        type: TypedEventAttributeType.URL,
-        value: URL.prototype.toString.call(value),
-      };
-    } else if (isString(value)) {
-      typedAttrValue = {
-        type: TypedEventAttributeType.String,
-        value,
-      };
-    } else if (isNumber(value)) {
-      typedAttrValue = {
-        type:
-          value % 1 === 0
-            ? TypedEventAttributeType.Integer
-            : TypedEventAttributeType.Float,
-        value,
-      };
-    } else if (isBoolean(value)) {
-      typedAttrValue = {
-        type: TypedEventAttributeType.Boolean,
-        value,
-      };
-    } else {
-      writeBatchLog(
-        false,
-        "BatchEventData - Invalid attribute value type. Must be a string, number, date, URL or boolean"
-      );
-      return this;
-    }
-
-    if (typedAttrValue) {
-      this._attributes[key] = typedAttrValue;
-    }
-
-    return this;
-  }
-
-  public _toInternalRepresentation(): IEventDataInternalRepresentation {
-    return {
-      attributes: this._attributes,
-      tags: Object.keys(this._tags),
-    };
-  }
-}
